csv_generator/requirements.txt
-i https://pypi.org/simple
boto3==1.12.12
botocore==1.15.12
decorator==4.4.2
docutils==0.15.2
jmespath==0.9.5
jsonpath-ng==1.4.3
ply==3.11
python-dateutil==2.8.1
python-json-logger==0.1.11
pyyaml==5.3
s3transfer==0.3.3
six==1.14.0
requests==2.7.0
pandas==1.1.3
numpy==1.16.4